pyproject.toml
[project]
name = "crafting"
version = "0.1.0"
description = ""
authors = [
    {name = "Valtteri Aurela",email = "valtteri@hamina.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.2.4,<3.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "pyqt5 (>=5.15.11,<6.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "albumentations (>=2.0.5,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "timm (>=1.0.15,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
